requirements.txt
alabaster==0.7.10
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
CacheControl==0.12.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
colorama==0.3.9
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
decorator==4.1.2
distlib==0.2.5
distributed==1.19.1
dj-database-url==0.5.0
Django==2.0.4
django-nose==1.4.5
djangorestframework==3.8.2
docutils==0.14
Jinja2==2.9.6
jsonschema==2.6.0
llvmlite==0.22.0
locket==0.2.0
lockfile==0.12.2
lxml==4.1.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.7.4
mpmath==0.19
multipledispatch==0.4.9
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nose==1.3.7
numpy
numpydoc==0.7.0
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
py==1.4.34
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
requests==2.18.4
rope==0.10.5
ruamel-yaml==0.13.11
scikit-learn==0.19.1
scipy==0.19.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
gunicorn
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.1
widgetsnbextension==3.0.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
django_heroku
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
zict==0.1.3
turicreate
django_heroku